Formatting and split auto-config tests.
Until we decided what to do with the reactive and imperative auto-configuration,
we will keep both. To remove the `@Nested` annotation but keep it readable the
tests get split up into different classes.
